requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
boto3==1.28.67
botocore==1.31.67
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chain==1.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.5.14
docutils==0.20.1
duckduckgo-search==3.9.3
extra-streamlit-components==0.1.60
Flask==3.0.0
Flask-Login==0.6.3
fonttools==4.43.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.320
langsmith==0.0.49
littleutils==0.2.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
mizani==0.9.3
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
outdated==0.2.2
packaging==23.1
pandas==2.1.0
pandas-flavor==0.6.0
patsy==0.5.3
Pillow==9.5.0
pingouin==0.5.3
plotly==5.18.0
plotly-express==0.4.1
plotnine==0.12.3
protobuf==4.24.3
psutil==5.9.6
psycopg2==2.9.9
pyarrow==13.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.3
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.20
statistics==1.0.3.5
statsmodels==0.14.0
streamlit==1.28.1
streamlit-cookies-manager==0.2.0
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trycourier==5.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.0
xarray==2023.10.1
yarl==1.9.2
zipp==3.16.2
